+++ app/src/main/java/camp/nextstep/edu/memo/main/MainViewModel.kt
@@ -1,6 +1,7 @@
 package camp.nextstep.edu.memo.main
 
 import androidx.lifecycle.ViewModel
+import camp.nextstep.edu.memo.MemoEvent
 import camp.nextstep.edu.memo.data.MemoRepositoryImpl
 import camp.nextstep.edu.memo.domain.entity.Memo
 import camp.nextstep.edu.memo.domain.repository.MemoRepository
@@ -15,7 +16,15 @@ class MainViewModel(
     private val _memoList = MutableStateFlow<List<Memo>>(emptyList())
     val memoList: StateFlow<List<Memo>> get() = _memoList.asStateFlow()
 
+    private val _memoEvent = MutableStateFlow<MemoEvent>(MemoEvent.None)
+    val memoEvent: StateFlow<MemoEvent> get() = _memoEvent.asStateFlow()
+
     fun fetchMemoList() {
         _memoList.value = memoRepository.fetch()
+    }
+
+    fun delete(position: Int) {
+        memoRepository.delete(position)
+        _memoEvent.value = MemoEvent.Delete(position)
     }
 }
