+++ app/src/test/java/camp/nextstep/edu/memo/main/MainViewModelTest.kt
@@ -1,10 +1,12 @@
 package camp.nextstep.edu.memo.main
 
+import camp.nextstep.edu.memo.MemoEvent
 import camp.nextstep.edu.memo.domain.entity.Memo
 import camp.nextstep.edu.memo.domain.repository.MemoRepository
 import camp.nextstep.edu.memo.main.MainViewModel
 import io.mockk.every
 import io.mockk.mockk
+import io.mockk.verify
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
@@ -33,5 +35,15 @@ internal class MainViewModelTest {
         viewModel.fetchMemoList()
 
         assertEquals(viewModel.memoList.value, memoList)
+    }
+
+    @Test
+    fun `특정 메모를 삭제합니다`() {
+        viewModel.delete(position = 0)
+
+        verify {
+            memoRepository.delete(position = 0)
+        }
+        assertEquals(viewModel.memoEvent.value, MemoEvent.Delete(position = 0))
     }
 }
